refactor(repositories): refactor update imports after changing the name of the package repositories to repository (#11)
